move geofabrik json class to own file
